requirements.txt
flask
pymupdf
sentence-transformers
faiss-cpu
numpy
